+++ src/FederationLib/Classes/Resources/operators.sql
@@ -11,7 +11,9 @@ create table operators
     created          timestamp   default current_timestamp() not null comment 'The Timestamp for when this operator record was created',
     updated          timestamp   default current_timestamp() not null comment 'The Timestamp for when this operator record was last updated',
     constraint operators_api_key_uindex
-        unique (api_key)
+        unique (api_key),
+    constraint operators_name_uindex
+        unique (name)
 );
 
 create index operators_created_index